// src/grpc/gen/nockchain/common/v1/primitives.ts
export interface Hash {
  belt_1?: Belt | undefined;
  belt_2?: Belt | undefined;
  belt_3?: Belt | undefined;
  belt_4?: Belt | undefined;
  belt_5?: Belt | undefined;
}
// ...
export interface Belt {
  value: string;
}
// ...
export const Hash: MessageFns<Hash> = {
  encode(message: Hash, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.belt_1 !== undefined) {
      Belt.encode(message.belt_1, writer.uint32(10).fork()).join();
    }
    if (message.belt_2 !== undefined) {
      Belt.encode(message.belt_2, writer.uint32(18).fork()).join();
    }
    if (message.belt_3 !== undefined) {
      Belt.encode(message.belt_3, writer.uint32(26).fork()).join();
    }
    if (message.belt_4 !== undefined) {
      Belt.encode(message.belt_4, writer.uint32(34).fork()).join();
    }
    if (message.belt_5 !== undefined) {
      Belt.encode(message.belt_5, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Hash {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseHash();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.belt_1 = Belt.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.belt_2 = Belt.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.belt_3 = Belt.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.belt_4 = Belt.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.belt_5 = Belt.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Hash {
    return {
      belt_1: isSet(object.belt1)
        ? Belt.fromJSON(object.belt1)
        : isSet(object.belt_1)
        ? Belt.fromJSON(object.belt_1)
        : undefined,
      belt_2: isSet(object.belt2)
        ? Belt.fromJSON(object.belt2)
        : isSet(object.belt_2)
        ? Belt.fromJSON(object.belt_2)
        : undefined,
      belt_3: isSet(object.belt3)
        ? Belt.fromJSON(object.belt3)
        : isSet(object.belt_3)
        ? Belt.fromJSON(object.belt_3)
        : undefined,
      belt_4: isSet(object.belt4)
        ? Belt.fromJSON(object.belt4)
        : isSet(object.belt_4)
        ? Belt.fromJSON(object.belt_4)
        : undefined,
      belt_5: isSet(object.belt5)
        ? Belt.fromJSON(object.belt5)
        : isSet(object.belt_5)
        ? Belt.fromJSON(object.belt_5)
        : undefined,
    };
  },

  toJSON(message: Hash): unknown {
    const obj: any = {};
    if (message.belt_1 !== undefined) {
      obj.belt1 = Belt.toJSON(message.belt_1);
    }
    if (message.belt_2 !== undefined) {
      obj.belt2 = Belt.toJSON(message.belt_2);
    }
    if (message.belt_3 !== undefined) {
      obj.belt3 = Belt.toJSON(message.belt_3);
    }
    if (message.belt_4 !== undefined) {
      obj.belt4 = Belt.toJSON(message.belt_4);
    }
    if (message.belt_5 !== undefined) {
      obj.belt5 = Belt.toJSON(message.belt_5);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Hash>, I>>(base?: I): Hash {
    return Hash.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Hash>, I>>(object: I): Hash {
    const message = createBaseHash();
    message.belt_1 = (object.belt_1 !== undefined && object.belt_1 !== null)
      ? Belt.fromPartial(object.belt_1)
      : undefined;
    message.belt_2 = (object.belt_2 !== undefined && object.belt_2 !== null)
      ? Belt.fromPartial(object.belt_2)
      : undefined;
    message.belt_3 = (object.belt_3 !== undefined && object.belt_3 !== null)
      ? Belt.fromPartial(object.belt_3)
      : undefined;
    message.belt_4 = (object.belt_4 !== undefined && object.belt_4 !== null)
      ? Belt.fromPartial(object.belt_4)
      : undefined;
    message.belt_5 = (object.belt_5 !== undefined && object.belt_5 !== null)
      ? Belt.fromPartial(object.belt_5)
      : undefined;
    return message;
  },
};
// ...
export const Belt: MessageFns<Belt> = {
  encode(message: Belt, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.value !== "0") {
      writer.uint32(8).uint64(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Belt {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBelt();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.value = reader.uint64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Belt {
    return { value: isSet(object.value) ? globalThis.String(object.value) : "0" };
  },

  toJSON(message: Belt): unknown {
    const obj: any = {};
    if (message.value !== "0") {
      obj.value = message.value;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Belt>, I>>(base?: I): Belt {
    return Belt.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Belt>, I>>(object: I): Belt {
    const message = createBaseBelt();
    message.value = object.value ?? "0";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}